fix: added image and refactor code
