+++ components/hero/index.ts
@@ -1,1 +1,3 @@
 export * from "./hero";
+export * from "./heroImage";
+export * from "./quotes";
